+++ src/StringEncoder/MB/Validator.php
@@ -11,23 +11,32 @@ class Validator
     /**
      * @internal
      */
-    public function validateEncoding(string $encoding): bool
+    public function validateEncoding(string $encoding, bool $caseSensitive): bool
+    {
+        return $this->determineEncoding($encoding, $caseSensitive) !== null;
+    }
+
+    /**
+     * @internal
+     */
+    public function determineEncoding(string $encoding, bool $caseSensitive): ?string
     {
         $encodingList = \mb_list_encodings();
-        $valid = false;
-        if (!\in_array($encoding, $encodingList)) {
-            // encoding not found in encoding list, checking aliases
-            foreach ($encodingList as $validEncoding) {
-                if ($this->validateEncodingAlias($encoding, $validEncoding)) {
-                    $valid = true;
-                    break;
-                }
+
+        foreach ($encodingList as $validEncoding) {
+            if ($validEncoding === $encoding || (
+                    $caseSensitive === false &&
+                    mb_convert_case($validEncoding, MB_CASE_LOWER) === mb_convert_case($encoding, MB_CASE_LOWER))
+                ) {
+                return $validEncoding;
             }
-        } else {
-            $valid = true;
+            if ($this->validateEncodingAlias($encoding, $validEncoding, $caseSensitive)) {
+                return $validEncoding;
+            }
         }
 
-        return $valid;
+        // no valid encoding string found that matches
+        return null;
     }
 
     public function validateString(string $string, EncodingDTOInterface $encodingDTO): bool
@@ -37,10 +46,33 @@ class Validator
         return $encoding === $encodingDTO->getEncoding();
     }
 
-    private function validateEncodingAlias(string $encoding, string $validEncoding): bool
+    /**
+     * @internal
+     */
+    private function validateEncodingAlias(string $encoding, string $validEncoding, bool $caseSensitive): bool
     {
-        $aliasEncoding = \mb_encoding_aliases($validEncoding);
+        if ($caseSensitive) {
+            $aliasEncoding = \mb_encoding_aliases($validEncoding);
+        } else {
+            $encoding = mb_convert_case($encoding, MB_CASE_LOWER);
+            $aliasEncoding = $this->lowerCaseArray(\mb_encoding_aliases($validEncoding));
+        }
 
         return \in_array($encoding, $aliasEncoding);
+    }
+
+    /**
+     * @internal
+     * @param string[] $encodings
+     * @return string[]
+     */
+    private function lowerCaseArray(array $encodings): array
+    {
+        $newEncodings = [];
+        foreach ($encodings as $key => $encoding) {
+            $newEncodings[$key] = mb_convert_case($encoding, MB_CASE_LOWER);
+        }
+
+        return $newEncodings;
     }
 }