+++ tests/tests/DTO/EncodingDTOTest.php
@@ -7,6 +7,7 @@ namespace tests\DTO;
 use PHPUnit\Framework\TestCase;
 use StringEncoder\DTO\EncodingDTO;
 use StringEncoder\Exceptions\InvalidEncodingException;
+use StringEncoder\Options;
 
 class EncodingDTOTest extends TestCase
 {
@@ -20,5 +21,11 @@ class EncodingDTOTest extends TestCase
     {
         $this->expectException(InvalidEncodingException::class);
         EncodingDTO::makeFromString('NOTANENCODING');
+    }
+
+    public function testMakeFromStringCaseInsensitive()
+    {
+        $encodingDTO = EncodingDTO::makeFromString('utf-8', null, (new Options())->setCaseSensitiveEncoding(false));
+        $this->assertEquals('UTF-8', $encodingDTO->getEncoding());
     }
 }