+++ tests/tests/MB/ValidatorTest.php
@@ -21,16 +21,31 @@ class ValidatorTest extends TestCase
 
     public function testValidateEncoding()
     {
-        $this->assertTrue($this->validator->validateEncoding('UTF-8'));
+        $this->assertTrue($this->validator->validateEncoding('UTF-8', true));
+    }
+
+    public function testValidateEncodingCaseInsensitive()
+    {
+        $this->assertTrue($this->validator->validateEncoding('utf-8', false));
     }
 
     public function testValidateEncodingInvalid()
     {
-        $this->assertFalse($this->validator->validateEncoding('NOTANENCODING'));
+        $this->assertFalse($this->validator->validateEncoding('NOTANENCODING', true));
+    }
+
+    public function testValidateEncodingInvaliCase()
+    {
+        $this->assertFalse($this->validator->validateEncoding('utf-8', true));
     }
 
     public function testValidateEncodingAlias()
     {
-        $this->assertTrue($this->validator->validateEncoding('US-ASCII'));
+        $this->assertTrue($this->validator->validateEncoding('US-ASCII', true));
+    }
+
+    public function testValidateEncodingAliasCaseInsensitive()
+    {
+        $this->assertTrue($this->validator->validateEncoding('us-ascII', false));
     }
 }